feat(backend): add jwt authentication

+++ backend/app/core/viewsets/employee.py
@@ -3,9 +3,12 @@ from rest_framework.status import (
     HTTP_201_CREATED,
     HTTP_204_NO_CONTENT,
     HTTP_400_BAD_REQUEST,
+    HTTP_403_FORBIDDEN,
 )
+from ..decorators.decode_jwt import require_auth
 from rest_framework.viewsets import ViewSet
 from rest_framework.response import Response
+from ..utils.auth import hash_password, create_access_token
 from ..domain.employee.employee_datastore import EmployeeDataStore
 
 
@@ -24,6 +27,7 @@ class Employee(ViewSet):
             "updated_at": employee_bean.updated_at,
         }
 
+    @require_auth
     def list(self, request, company_id=None):
         if not company_id:
             return Response(
@@ -33,6 +37,7 @@ class Employee(ViewSet):
         response_data = [self._bean_to_dict(bean) for bean in beans]
         return Response(response_data, status=HTTP_200_OK)
 
+    @require_auth
     def retrieve(self, request, company_id=None, pk=None):
         if not pk or not company_id:
             return Response(
@@ -44,15 +49,34 @@ class Employee(ViewSet):
         return Response(response_data, status=HTTP_200_OK)
 
     def create(self, request, company_id=None):
-        bean = EmployeeDataStore.create_employee(company_id, request.data)
-        response_data = self._bean_to_dict(bean)
+        data = request.data
+        data["password"] = hash_password(data["password"])
+        bean = EmployeeDataStore.create_employee(company_id, data)
+        response_data = {
+            "employee": self._bean_to_dict(bean),
+            "access_token": create_access_token(bean.id),
+        }
         return Response(response_data, status=HTTP_201_CREATED)
 
+    @require_auth
     def destroy(self, request, company_id=None, pk=None):
+        if str(request.user_id) != str(pk):
+            return Response(
+                {"error": "You can only delete your own account"},
+                status=HTTP_403_FORBIDDEN,
+            )
+
         EmployeeDataStore.delete_employee(pk)
         return Response(status=HTTP_204_NO_CONTENT)
 
+    @require_auth
     def update(self, request, company_id=None, pk=None):
+        if str(request.user_id) != str(pk):
+            return Response(
+                {"error": "You can only delete your own account"},
+                status=HTTP_403_FORBIDDEN,
+            )
+
         bean = EmployeeDataStore.update_employee(pk, request.data)
         response_data = self._bean_to_dict(bean)
         return Response(response_data, status=HTTP_200_OK)
